refactor: 유틸리티 함수 재구성
- 여러 컴포넌트에서 동시에 사용하는 경우가 아니라면 사용되는 단일
  컴포넌트 또는 해당 컴포넌트가 위치한 모듈 내로 위치 변경
