+++ src/components/common/toast.tsx
@@ -3,8 +3,9 @@
 import { ToastType } from '@/utils/type';
 import { useCountdown } from '@/hooks/useCountdown';
 import { useEffect } from 'react';
-import { TOAST_VISIBILITY_DURATION_SECOND } from '@/utils/constant';
 import { toastStyle } from '@/styles/common/toastStyle.css';
+
+const TOAST_VISIBILITY_DURATION_SECOND = 5;
 
 interface ToastProps {
   message: string;
