+++ src/components/video/videoPlayerControlPanel.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { ChangeEvent, Dispatch, RefObject, SetStateAction, useRef } from 'react';
-import { numberToTime } from '@/utils/time';
 import { throttle } from '@/utils/throttle';
 import Slider from '../common/slider';
 import { setMuteStorageValue, setVolumeStorageValue } from '@/utils/storage';
@@ -9,6 +8,7 @@ import * as Popover from '@radix-ui/react-popover';
 import { EnterFullScreenIcon, ExitFullScreenIcon, PauseIcon, PlayIcon, SpeakerLoudIcon, SpeakerOffIcon } from '@radix-ui/react-icons';
 import { VideoResolutionLevel } from './videoPlayer';
 import { videoPlayerControlPanelStyle } from '@/styles/video/videoPlayerControlPanelStyle.css';
+import { numberToTime } from '@/utils/convert';
 
 interface VideoPlayerControlPanelProps {
   videoRef: RefObject<HTMLVideoElement | null>;
